ordinarias delete fix

+++ codigoFuente/basegestor/src/main/java/mx/gob/sedesol/basegestor/model/repositories/gestionescolar/InscripcionesRepository.java
@@ -319,7 +319,7 @@ public class InscripcionesRepository implements IinscripcionesRepository {
 	
 	@Transactional
 	@Override
-	public void deleteProcesoInscripcion(Long procesoInscripcionId, Long convocatoriaId) {
+	public void deleteProcesoInscripcion(Long procesoInscripcionId, Long convocatoriaId, String tipoProceso) {
 	    try {
 	        // 1. Eliminar registros relacionados en tbl_terminosycondiciones
 	        String deleteTerminosQuery = "DELETE FROM tbl_terminosycondiciones WHERE id_proceso_inscripcion = ?";
@@ -337,19 +337,22 @@ public class InscripcionesRepository implements IinscripcionesRepository {
 	        String deleteDispersionesQuery = "DELETE FROM tbl_dispersiones WHERE id_proceso_inscripcion = ?";
 	        entityManager.createNativeQuery(deleteDispersionesQuery)
 	                     .setParameter(1, procesoInscripcionId)
-	                     .executeUpdate();
-
-	        // 4. Eliminar registros relacionados en rel_proceso_inscipcion_planesyprogramas
-	        String deleteRelacionesQuery = "DELETE FROM rel_proceso_inscipcion_planesyprogramas WHERE id_proceso_inscripcion = ?";
-	        entityManager.createNativeQuery(deleteRelacionesQuery)
-	                     .setParameter(1, procesoInscripcionId)
-	                     .executeUpdate();
+	                     .executeUpdate();    
+            //4. Eliminar registros en rel_proceso_inscipcion_planesyprogramas
+            String deleteRelacionesQuery = "DELETE FROM rel_proceso_inscipcion_planesyprogramas WHERE id_proceso_inscripcion = ?";
+            entityManager.createNativeQuery(deleteRelacionesQuery)
+                                                 .setParameter(1, procesoInscripcionId)
+                                                 .executeUpdate();
 	        
-	        // 5. Eliminar registros relacionados en rel_proceso_inscipcion_planesyprogramas
-	        String deleteResumenQuery = "DELETE FROM tbl_inscripcion_resumen WHERE id_convocatoria = ?";
-	        entityManager.createNativeQuery(deleteResumenQuery)
-	                     .setParameter(1, convocatoriaId)
-	                     .executeUpdate();
+	        if (tipoProceso.equalsIgnoreCase("Ordinario")) {
+	        	
+	            // 5. Eliminar registros en tbl_inscripcion_resumen
+		        String deleteResumenQuery = "DELETE FROM tbl_inscripcion_resumen WHERE id_convocatoria = ?";
+		        entityManager.createNativeQuery(deleteResumenQuery)
+		                     .setParameter(1, convocatoriaId)
+		                     .executeUpdate();
+
+	        }
 	        
 	        // 6. Eliminar el registro principal en tbl_procesos_inscripcion
 	        String deletePrincipalQuery = "DELETE FROM tbl_procesos_inscripcion WHERE proceso_inscripcion_id = ?";
@@ -360,6 +363,7 @@ public class InscripcionesRepository implements IinscripcionesRepository {
 	        throw new RuntimeException("Error al eliminar el proceso de inscripción y sus relaciones", e);
 	    }
 	}
+	
 //	
 //	@Override
 //	public List<ConvocatoriaTableroResumen> consultarTableroResumen(Integer convocatoriaId) {
